requirements.txt
alpaca-py>=0.13.0
python-dotenv>=0.19.0
pandas>=1.3.0
numpy>=1.21.0
schedule>=1.1.0
pytz>=2021.3
holidays>=0.14.0
